move add_plot_data_to_plot_collection to util module
